Move validation.ts and jsonc.ts from config directory to lib directory
