+++ src/lib/jsonc.ts
